# cache: write blob to the content store via bufio during conversion