--- a/cache/converter.go
+++ b/cache/converter.go
@@ -1,9 +1,11 @@
 package cache
 
 import (
+	"bufio"
 	"context"
 	"fmt"
 	"io"
+	"sync"
 
 	cdcompression "github.com/containerd/containerd/archive/compression"
 	"github.com/containerd/containerd/content"
@@ -120,6 +122,12 @@ type conversion struct {
 	finalize   func(context.Context, content.Store) (map[string]string, error)
 }
 
+var bufioPool = sync.Pool{
+	New: func() interface{} {
+		return nil
+	},
+}
+
 func (c *conversion) convert(ctx context.Context, cs content.Store, desc ocispecs.Descriptor) (*ocispecs.Descriptor, error) {
 	// prepare the source and destination
 	labelz := make(map[string]string)
@@ -132,16 +140,24 @@ func (c *conversion) convert(ctx context.Context, cs content.Store, desc ocispec
 	if err := w.Truncate(0); err != nil { // Old written data possibly remains
 		return nil, err
 	}
-	var zw io.WriteCloser = w
-	var compress io.WriteCloser
+
+	var bufW *bufio.Writer
+	if pooledW := bufioPool.Get(); pooledW != nil {
+		bufW = pooledW.(*bufio.Writer)
+		bufW.Reset(w)
+	} else {
+		bufW = bufio.NewWriterSize(w, 128*1024)
+	}
+	defer bufioPool.Put(bufW)
+	var zw io.WriteCloser = &nopWriteCloser{bufW}
 	if c.compress != nil {
 		zw, err = c.compress(zw)
 		if err != nil {
 			return nil, err
 		}
-		defer zw.Close()
-		compress = zw
 	}
+	zw = &onceWriteCloser{WriteCloser: zw}
+	defer zw.Close()
 
 	// convert this layer
 	diffID := digest.Canonical.Digester()
@@ -164,10 +180,11 @@ func (c *conversion) convert(ctx context.Context, cs content.Store, desc ocispec
 	if _, err := io.Copy(zw, io.TeeReader(rdr, diffID.Hash())); err != nil {
 		return nil, err
 	}
-	if compress != nil {
-		if err := compress.Close(); err != nil { // Flush the writer
-			return nil, err
-		}
+	if err := zw.Close(); err != nil { // Flush the writer
+		return nil, err
+	}
+	if err := bufW.Flush(); err != nil { // Flush the buffer
+		return nil, errors.Wrap(err, "failed to flush diff during conversion")
 	}
 	labelz[labels.LabelUncompressed] = diffID.Digest().String() // update diffID label
 	if err = w.Commit(ctx, 0, "", content.WithLabels(labelz)); err != nil && !errdefs.IsAlreadyExists(err) {
@@ -210,4 +227,24 @@ func (rc *readCloser) Close() error {
 		return errors.Wrapf(err1, "failed to close: %v", err2)
 	}
 	return err2
+}
+
+type nopWriteCloser struct {
+	io.Writer
+}
+
+func (w *nopWriteCloser) Close() error {
+	return nil
+}
+
+type onceWriteCloser struct {
+	io.WriteCloser
+	closeOnce sync.Once
+}
+
+func (w *onceWriteCloser) Close() (err error) {
+	w.closeOnce.Do(func() {
+		err = w.WriteCloser.Close()
+	})
+	return
 }